requirements.txt
colorama
cryptography
dnspython
publicsuffixlist